Refactor ZoneDistribution and clean up heart rate zone logic
