+++ components/index.ts
@@ -6,7 +6,6 @@ export * from './Footer'
 export * from './FooterControls'
 export * from './GoogleDocViewer'
 export * from './HRMonitorStatusIndicator'
-export * from './HeartRateZones'
 export * from './HrTile'
 export * from './HrmConnectionPanel'
 export * from './HrmTiles'
