+++ theme/theme.ts
@@ -3,47 +3,6 @@
 import { createTheme } from '@mui/material/styles'
 import libTheme from '@/lib/theme'
 
-declare module '@mui/material/styles' {
-  interface Palette {
-    custom: {
-      prepare: string
-      work: string
-      rest: string
-      running: string
-      idle: string
-      cooldown: string
-      hrZones: {
-        warmUp: string
-        fatBurn: string
-        cardio: string
-        peak: string
-        max: string
-        unknown: string
-        noData: string
-      }
-    }
-  }
-  interface PaletteOptions {
-    custom?: {
-      prepare?: string
-      work?: string
-      rest?: string
-      running?: string
-      idle?: string
-      cooldown?: string
-      hrZones?: {
-        warmUp?: string
-        fatBurn?: string
-        cardio?: string
-        peak?: string
-        max?: string
-        unknown?: string
-        noData?: string
-      }
-    }
-  }
-}
-
 const theme = createTheme(libTheme, {
   palette: {
     custom: {
