+++ types/mui-theme.d.ts
@@ -1,0 +1,42 @@
+import '@mui/material/styles'
+
+declare module '@mui/material/styles' {
+  interface Palette {
+    custom: {
+      prepare: string
+      work: string
+      rest: string
+      running: string
+      idle: string
+      cooldown: string
+      hrZones: {
+        warmUp: string
+        fatBurn: string
+        cardio: string
+        peak: string
+        max: string
+        unknown: string
+        noData: string
+      }
+    }
+  }
+  interface PaletteOptions {
+    custom?: {
+      prepare?: string
+      work?: string
+      rest?: string
+      running?: string
+      idle?: string
+      cooldown?: string
+      hrZones?: {
+        warmUp?: string
+        fatBurn?: string
+        cardio?: string
+        peak?: string
+        max?: string
+        unknown?: string
+        noData?: string
+      }
+    }
+  }
+}
